copyExampleFiles {spatstat.data}	R Documentation

Copy Data Files for Example

Description

This command copies several data files to a folder (directory) chosen by the user, so that they can be used for a practice example.

Usage

copyExampleFiles(which, folder = getwd())

Arguments

which
Character string name (partially matched) of one of the datasets installed in spatstat for which the original data files are provided. If which is missing, a list of available options is printed.

folder
Character string path name of a folder (directory) in which the files will be placed. Defaults to the current working directory.

Details

The original text files containing data for the selected dataset are copied to the chosen folder.

This is part of an exercise described in Chapter 3 of Baddeley, Rubak and Turner (2015).

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

References

Baddeley, A., Rubak, E. and Turner, R. (2015) Spatial Point Patterns: Methodology and Applications with R. Chapman and Hall/CRC Press.

See Also

vesicles, finpines

Examples

   copyExampleFiles()

[Package spatstat.data version 3.1-4 Index]