demohyper {spatstat.data}	R Documentation

Demonstration Example of Hyperframe of Spatial Data

Description

This is an artificially constructed example of a hyperframe of spatial data. The data could have been obtained from an experiment in which there are two groups of experimental units, the response from each unit is a point pattern Points, and for each unit there is explanatory data in the form of a pixel image Image.

Usage

data(demohyper)

Format

A hyperframe with 3 rows and 3 columns:

Points

List of spatial point patterns (objects of class "ppp") serving as the responses in an experiment.

Image

List of images (objects of class "im") serving as explanatory variables.

Group

Factor with two levels a and b serving as an explanatory variable.

Source

Artificially generated by Adrian Baddeley Adrian.Baddeley@curtin.edu.au.

Examples

  if(require(spatstat.model)) {
 plot(demohyper, quote({ plot(Image, main=""); plot(Points, add=TRUE) }),
      parargs=list(mar=rep(1,4)))
 mppm(Points ~ Group/Image, data=demohyper)
  }

[Package spatstat.data version 3.1-4 Index]