demopat {spatstat.data}	R Documentation

Artificial Data Point Pattern

Description

This is an artificial dataset, for use in testing and demonstrating the capabilities of the spatstat package.

It is a multitype point pattern in an irregular polygonal window. There are two types of points. The window contains a polygonal hole. Spatial coordinates are expressed in furlongs (one furlong equals 660 feet).

Usage

data(demopat)

Format

An object of class "ppp" representing the point pattern.

See ppp.object for details of the format of a point pattern object.

Source

Adrian Baddeley Adrian.Baddeley@curtin.edu.au

[Package spatstat.data version 3.1-4 Index]