flu {spatstat.data}	R Documentation

Influenza Virus Proteins

Description

Replicated spatial point patterns giving the locations of two different virus proteins on the membranes of cells infected with influenza virus.

Usage

data(flu)

Format

A hyperframe with 41 rows and four columns:

pattern

List of spatial point patterns (objects of class "ppp") with points of two types, identifying the locations of two different proteins on a membrane sheet. Coordinates are expressed in nanometres (nm) and the window of observation is a square of side length 3331 nm.

virustype

Factor identifying whether the infecting virus was the wild type (wt) or mutant (mut1).

stain

Factor identifying whether the membrane sheet was stained for the proteins M2 and M1 (stain="M2-M1") or stained for the proteins M2 and HA (stain="M2-HA").

frameid

Integer. Serial number of the microscope frame in the original experiment. Frame identifier is not unique across different values of virustype and stain.

The row names of the hyperframe can be used as succinct labels in plots.

Details

The data consist of 41 spatial point patterns, each giving the locations of two different virus proteins on the membranes of cells infected with influenza virus.

Chen et al (2008) conducted the experiment and used spatial analysis to establish evidence for an interaction between the influenza virus proteins M1 and M2 that is important for the study of viral replication.

Canine kidney cells were infected with human influenza, Udorn strain, either the wild type or a mutant which encodes a defective M2 protein. At twelve hours post-infection, membrane sheets were prepared and stained for viral proteins, using two antibodies conjugated to gold particles of two sizes (6 nanometre and 12 nanometre diameter) enabling localisation of two different proteins on each sheet. The 6 nm particles were stained for M2 (ion channel protein), while the 12 nm particles were stained either for M1 (matrix protein) or for HA (hemagglutinin). Membrane sheets were visualised in electron microscopy.

Experimental technique and spatial analysis of the membranes stained for M2 and M1 is reported in Chen et al (2008). Analysis of the membranes stained for M2 and HA is reported in Rossman et al (2010). The M2-HA data shows a stronger association between the two proteins which has also been observed biochemically and functionally (Rossman et al, 2010).

The dataset flu is a hyperframe with one row for each membrane sheet. The column named pattern contains the spatial point patterns of gold particle locations, with two types of points (either M1 and M2 or HA and M2). The column named virustype is a factor identifying the virus: either wild type wt or mutant mut1. The column named stain is a factor identifying whether the membrane was stained for M1 and M2 (stain="M2-M1") or stained for HA and M2 (stain="M2-HA"). The row names of the hyperframe are a succinct summary of the experimental conditions and can be used as labels in plots. See the Examples.

Source

Data generously provided by Dr G.P. Leser and Dr R.A. Lamb. Please cite Chen et al (2008) in any use of these data.

References

Chen, B.J., Leser, G.P., Jackson, D. and Lamb, R.A. (2008) The influenza virus M2 protein cytoplasmic tail interacts with the M1 protein and influences virus assembly at the site of virus budding. Journal of Virology 82, 10059–10070.

Rossman, J.S., Jing, X.H., Leser, G.P. and Lamb, R.A. (2010) Influenza virus M2 protein mediates ESCRT-independent membrane scission Cell 142, 902–913.

Examples

  if(require(spatstat.geom)) {
flu
Y <- flu$pattern[10]
Y <- flu[10, 1, drop=TRUE]
wildM1 <- with(flu, virustype == "wt" & stain == "M2-M1")
plot(flu[wildM1, 1, drop=TRUE], 
     main=c("flu data", "wild type virus, M2-M1 stain"),
     pch=c(3,16), cex=0.4, cols=2:3)
  }

[Package spatstat.data version 3.1-4 Index]