simba {spatstat.data}	R Documentation

Simulated data from a two-group experiment with replication within each group.

Description

The simba dataset contains simulated data from an experiment with a ‘control’ group and a ‘treatment’ group, each group containing 5 experimental units.

The responses in the experiment are point patterns.

The responses in the control group are independent realisations of a Poisson point process with intensity 80.

The responses in the treatment group are independent realisations of a Strauss process with activity parameter beta=100, interaction parameter gamma=0.5 and interaction radius R=0.07 in the unit square.

Usage

data(simba)

Format

simba is a hyperframe with 10 rows, and columns named:

Points containing the point patterns

group factor identifying the experimental group, with levels control and treatment).

Source

Simulated data, generated by Adrian Baddeley Adrian.Baddeley@curtin.edu.au.

[Package spatstat.data version 3.1-4 Index]