spatstat.data-package {spatstat.data}	R Documentation

The spatstat.data Package

Description

The spatstat.data package contains the datasets for the spatstat family of packages.

Details

The spatstat.data package contains the datasets for the spatstat family of packages.

These are spatial datasets; they are objects belonging to classes of spatial data defined in other sub-packages of the spatstat family. In order to handle these datasets correctly, we recommend loading the spatstat package.

Licence

This library and its documentation are usable under the terms of the “GNU General Public License”, a copy of which is distributed with R.

Author(s)

Adrian Baddeley Adrian.Baddeley@curtin.edu.au, Rolf Turner rolfturner@posteo.net and Ege Rubak rubak@math.aau.dk.

[Package spatstat.data version 3.1-4 Index]