spiders {spatstat.data}	R Documentation

Spider Webs on Mortar Lines of a Brick Wall

Description

Data recording the locations of small spider webs on the network of mortar lines of a brick wall.

Usage

data("spiders")

Format

Object of class "lpp" representing a pattern of points on a linear network. Spatial coordinates are expressed in millimetres.

Details

The data give the positions of 48 webs of the urban wall spider Oecobius navus on the mortar lines of a brick wall, recorded by Voss (1999) and manually digitised by Mark Handcock. The mortar spaces provide the only opportunity for constructing webs (Voss 1999; Voss et al 2007) so this is a pattern of points on a network of lines.

The habitat preferences of this species were studied in detail by Voss et al (2007). Questions of interest include evidence for non-uniform density of webs and for interaction between nearby individuals.

Observations were made inside a square quadrat of side length 1.125 metres. The original hand-drawn map was digitised manually by Mark S. Handcock, and reformatted as a spatstat object by Ang Qi Wei.

The dataset spiders is an object of class "lpp" (point pattern on a linear network). Coordinates are given in millimetres. The linear network has 156 vertices and a total length of 20.22 metres.

Please cite Voss et al (2007) with any use of these data.

Source

Dr Sasha Voss. Coordinates manually recorded by M.S. Handcock and formatted by Q.W. Ang.

Please cite Voss et al (2007) with any use of these data.

References

Ang, Q.W. (2010) Statistical methodology for events on a network. Master's thesis, School of Mathematics and Statistics, University of Western Australia.

Voss, S. (1999) Habitat preferences and spatial dynamics of the urban wall spider: Oecobius annulipes Lucas. Honours Thesis, Department of Zoology, University of Western Australia.

Voss, S., Main, B.Y. and Dadour, I.R. (2007) Habitat preferences of the urban wall spider Oecobius navus (Araneae, Oecobiidae). Australian Journal of Entomology 46, 261–268.

Examples

  if(require(spatstat.linnet)) {
plot(spiders, show.window=FALSE, pch=16)
   }

[Package spatstat.data version 3.1-4 Index]