spruces {spatstat.data}	R Documentation

Spruces Point Pattern

Description

The data give the locations of Norwegian spruce trees in a natural forest stand in Saxonia, Germany. Each tree is marked with its diameter at breast height.

Usage

data(spruces)

Format

An object of class "ppp" representing the point pattern of tree locations in a 56 x 38 metre sampling region. Each tree is marked with its diameter at breast height. All values are given in metres.

See ppp.object for details of the format of a point pattern object. The marks are numeric.

These data have been analysed by Fiksel (1984, 1988), Stoyan et al (1987), Penttinen et al (1992) and Goulard et al (1996).

Source

Stoyan et al (1987). Original source unknown.

References

Fiksel, T. (1984) Estimation of parameterized pair potentials of marked and nonmarked Gibbsian point processes. Elektron. Informationsverarb. u. Kybernet. 20, 270–278.

Fiksel, T. (1988) Estimation of interaction potentials of Gibbsian point processes. Statistics 19, 77-86

Goulard, M., S\"arkk\"a, A. and Grabarnik, P. (1996) Parameter estimation for marked Gibbs point processes through the maximum pseudolikelihood method. Scandinavian Journal of Statistics 23, 365–379.

Penttinen, A., Stoyan, D. and Henttonen, H. (1992) Marked point processes in forest statistics. Forest Science 38, 806–824.

Stoyan, D., Kendall, W.S. and Mecke, J. (1987) Stochastic Geometry and its Applications. Wiley.

Examples

  if(require(spatstat.geom)) {
     plot(spruces)
     # To reproduce Goulard et al. Figure 3
     # (Goulard et al: "influence zone radius equals 5 * stem diameter")
     # (help(plot.ppp) says: "size of symbol = diameter")
     plot(spruces, maxsize=10*max(spruces$marks))
     plot(unmark(spruces), add=TRUE)
  }

[Package spatstat.data version 3.1-4 Index]