urkiola {spatstat.data}	R Documentation

Urkiola Woods Point Pattern

Description

Locations of birch (Betula celtiberica) and oak (Quercus robur) trees in a secondary wood in Urkiola Natural Park (Basque country, northern Spain). They are part of a more extensive dataset collected and analysed by Laskurain (2008). The coordinates of the trees are given in meters.

Usage

data(urkiola)

Format

An object of class "ppp" representing the point pattern of tree locations. Entries include

x

Cartesian x-coordinate of tree in metres

y

Cartesian y-coordinate of tree in metres

marks

factor indicating species of each tree

The levels of marks are birch and oak. See ppp.object for details of the format of a ppp object.

Source

N.A. Laskurain. Kindly formatted and communicated by M. de la Cruz Rot

References

Laskurain, N. A. (2008) Dinámica espacio-temporal de un bosque secundario en el Parque Natural de Urkiola (Bizkaia). Tesis Doctoral. Universidad del País Vasco /Euskal Herriko Unibertsitatea.

[Package spatstat.data version 3.1-4 Index]